Move symbolToName into TreeSitter.Symbol.
