+++ src/TreeSitter/Symbol.hs
@@ -1,8 +1,11 @@
 {-# LANGUAGE ScopedTypeVariables #-}
 module TreeSitter.Symbol where
 
-import Data.Ix
-import Data.Word
+import Data.Char (isAlpha, toUpper)
+import Data.Function ((&))
+import Data.Ix (Ix)
+import Data.List.Split (condense, split, whenElt)
+import Data.Word (Word16)
 
 type TSSymbol = Word16
 
@@ -15,3 +18,62 @@ data SymbolType = Regular | Anonymous | Auxiliary
 
 class (Bounded s, Enum s, Ix s, Ord s, Show s) => Symbol s where
   symbolType :: s -> SymbolType
+
+
+symbolToName :: SymbolType -> String -> (SymbolType, String)
+symbolToName ty name
+  = prefixHidden name
+  & toWords
+  & filter (not . all (== '_'))
+  & map (>>= toDescription)
+  & (>>= initUpper)
+  & (prefix ++)
+  & (,) ty
+  where toWords = split (condense (whenElt (not . isAlpha)))
+
+        prefixHidden s@('_':_) = "Hidden" ++ s
+        prefixHidden s = s
+
+        initUpper (c:cs) = toUpper c : cs
+        initUpper "" = ""
+
+        toDescription '{' = "LBrace"
+        toDescription '}' = "RBrace"
+        toDescription '(' = "LParen"
+        toDescription ')' = "RParen"
+        toDescription '.' = "Dot"
+        toDescription ':' = "Colon"
+        toDescription ',' = "Comma"
+        toDescription '|' = "Pipe"
+        toDescription ';' = "Semicolon"
+        toDescription '*' = "Star"
+        toDescription '&' = "Ampersand"
+        toDescription '=' = "Equal"
+        toDescription '<' = "LAngle"
+        toDescription '>' = "RAngle"
+        toDescription '[' = "LBracket"
+        toDescription ']' = "RBracket"
+        toDescription '+' = "Plus"
+        toDescription '-' = "Minus"
+        toDescription '/' = "Slash"
+        toDescription '\\' = "Backslash"
+        toDescription '^' = "Caret"
+        toDescription '!' = "Bang"
+        toDescription '%' = "Percent"
+        toDescription '@' = "At"
+        toDescription '~' = "Tilde"
+        toDescription '?' = "Question"
+        toDescription '`' = "Backtick"
+        toDescription '#' = "Hash"
+        toDescription '$' = "Dollar"
+        toDescription '"' = "DQuote"
+        toDescription '\'' = "SQuote"
+        toDescription '\t' = "Tab"
+        toDescription '\n' = "LF"
+        toDescription '\r' = "CR"
+        toDescription c = [c]
+
+        prefix = case ty of
+          Regular -> ""
+          Anonymous -> "Anon"
+          Auxiliary -> "Aux"
